+++ src/App.tsx
@@ -8,6 +8,7 @@ import SettingsPanel from './components/SettingsPanel/SettingsPanel'
 import EmptyState from './components/EmptyState'
 import { useStore } from './store'
 import FloatingKeyboard from './components/Keyboard/FloatingKeyboard'
+import ChordExplorer from './components/ChordExplorer'
 import MidiEditor from './components/MidiEditor/MidiEditor'
 import { parseMidiBuffer } from './utils/midiParser'
 import { detectKeyFromTracks, parseKeySignature } from './utils/keyDetection'
@@ -50,10 +51,12 @@ export default function App() {
       switch (e.key) {
         case ' ':
           e.preventDefault()
+          if (useStore.getState().chordExplorerOpen) break
           if (playbackState === 'playing') pause()
           else play()
           break
         case 'Escape':
+          if (useStore.getState().chordExplorerOpen) break
           stop()
           break
         case 'o':
@@ -83,6 +86,7 @@ export default function App() {
         <TrackPanel />
       </div>
       {keyboardMode === 'floating' && <FloatingKeyboard />}
+      <ChordExplorer />
     </div>
   )
 }